requirements.txt
google-generativeai
PyGithub